// packages/build/src/types.ts
export interface ManifestEntry {
  integrity?: string
  revision: string | null
  url: string
}
// ...
export type ManifestTransform = (
  manifestEntries: Array<ManifestEntry & { size: number }>,
) => Promise<ManifestTransformResult> | ManifestTransformResult

export interface BasePartial {
  /**
   * A list of entries to be precached, in addition to any entries that are
   * generated as part of the build configuration.
   */
  additionalManifestEntries?: Array<string | ManifestEntry>
  /**
   * Async generator that yields additional entries to be preached.
   * runs **after** `additionalManifestEntries` and glob patterns, but **before** `manifestTransforms`
   * If the same URL appears more than once (e.g. from both static array and generator),
   * a warning will be logged and the duplicate will be skipped.
   */
  additionalManifestEntriesGenerator?: () => AsyncGenerator<string | ManifestEntry, undefined, void>
  /**
   * Assets that match this will be assumed to be uniquely versioned via their
   * URL, and exempted from the normal HTTP cache-busting that's done when
   * populating the precache. While not required, it's recommended that if your
   * existing build process already inserts a `[hash]` value into each filename,
   * you provide a RegExp that will detect that, as it will reduce the bandwidth
   * consumed when precaching.
   */
  dontCacheBustURLsMatching?: RegExp
  /**
   * One or more functions which will be applied sequentially against the
   * generated manifest. If `modifyURLPrefix` or `dontCacheBustURLsMatching` are
   * also specified, their corresponding transformations will be applied first.
   */
  manifestTransforms?: Array<ManifestTransform>
  /**
   * This value can be used to determine the maximum size of files that will be
   * precached. This prevents you from inadvertently precaching very large files
   * that might have accidentally matched one of your patterns.
   * @default 2097152
   */
  maximumFileSizeToCacheInBytes?: number
  /**
   * Should `maximumFileSizeToCacheInBytes` exceeded throw an error?.
   * @default true
   */
  throwMaximumFileSizeToCacheInBytes?: boolean

  /**
   * An object mapping string prefixes to replacement string values. This can be
   * used to, e.g., remove or add a path prefix from a manifest entry if your
   * web hosting setup doesn't match your local filesystem setup. As an
   * alternative with more flexibility, you can use the `manifestTransforms`
   * option and provide a function that modifies the entries in the manifest
   * using whatever logic you provide.
   *
   * Example usage:
   *
   * ```
   * // Replace a '/dist/' prefix with '/', and also prepend
   * // '/static' to every URL.
   * modifyURLPrefix: {
   *   '/dist/': '/',
   *   '': '/static',
   * }
   * ```
   */
  modifyURLPrefix?: {
    [key: string]: string
  }
  /**
   * When true, throw an error listing all duplicate precache entry URLs.
   * @default false
   */
  failOnDuplicateManifestEntries?: boolean
}

export type SWType = 'classic' | 'module' | 'classic-and-module'
// ...
export interface RequiredGlobDirectoryPartial {
  /**
   * The local directory you wish to match `globPatterns` against. The path is
   * relative to the current directory.
   */
  globDirectory: string
}

export interface OptionalGlobDirectoryPartial {
  /**
   * The local directory you wish to match `globPatterns` against. The path is
   * relative to the current directory.
   */
  globDirectory?: string
}

export interface GlobPartial {
  /**
   * Determines whether or not symlinks are followed when generating the
   * precache manifest. For more information, see the definition of `followSymbolicLinks`
   * in the `glob` [documentation](https://superchupu.dev/tinyglobby).
   * @default true
   */
  globFollow?: boolean
  /**
   * A set of patterns matching files to always exclude when generating the
   * precache manifest. For more information, see the definition of `ignore` in
   * the `glob` [documentation](https://superchupu.dev/tinyglobby).
   */
  globIgnores?: Array<string>
  /**
   * Files matching any of these patterns will be included in the precache
   * manifest. For more information, see the
   * [tinyglobby](https://superchupu.dev/tinyglobby).
   */
  globPatterns?: Array<string>
  /**
   * If true, an error reading a directory when generating a precache manifest
   * will cause the build to fail. If false, the problematic directory will be
   * skipped. For more information, see the definition of `strict` in the `glob`
   * [documentation](https://superchupu.dev/tinyglobby).
   * @default true
   */
  globStrict?: boolean
  /**
   * If a URL is rendered based on some server-side logic, its contents may
   * depend on multiple files or on some other unique string value. The keys in
   * this object are server-rendered URLs. If the values are an array of
   * strings, they will be interpreted as `glob` patterns, and the contents of
   * any files matching the patterns will be used to uniquely version the URL.
   * If used with a single string, it will be interpreted as unique versioning
   * information that you've generated for a given URL.
   */
  templatedURLs?: {
    [key: string]: string | Array<string>
  }
}

export interface InjectPartial {
  /**
   * @default 'info'
   */
  logLevel?: LogLevel
  /**
   * Granular control for underlying bundler `logLevel`.
   *
   * If you set `logLevel` to '`silent` then the underline bundler logLevel will also be `silent`.
   *
   * @default `{ rolldown: 'warn', vite: 'warn' }`
   */
  bundlerLogLevel?: BundlerLogLevel
  /**
   * The string to find inside of the `swSrc` file. Once found, it will be
   * replaced by the generated precache manifest.
   * Set it to `null` or `false` to disable injection point.
   *
   * **NOTE**: calling `injectManifest` directly with `injectionPoint` set to `null` or `false` will fail.
   *
   * @default "self.__WB_MANIFEST"
   */
  injectionPoint?: string | null | false
  /**
   * The path and filename of the service worker file that will be read during
   * the build process, relative to the current working directory.
   */
  swSrc: string
}

export interface RequiredSWDestPartial {
  /**
   * The path and filename of the service worker file that will be created by
   * the build process, relative to the current working directory. It must end
   * in '.js'.
   */
  swDest: string
}

export type GenerateSWOptions<T extends SWType> = BasePartial
  & GlobPartial
  & GeneratePartial<T>
  & RequiredSWDestPartial
  & OptionalGlobDirectoryPartial

export type GetManifestOptions = BasePartial
  & GlobPartial
  & RequiredGlobDirectoryPartial

export type InjectManifestOptions = BasePartial
  & GlobPartial
  & InjectPartial
  & RequiredSWDestPartial
  & RequiredGlobDirectoryPartial